requirements.txt
uvicorn==0.18.3
databases==0.6.1
python-decouple==3.6
SQLAlchemy~=1.4.40
fastapi~=0.81.0
requests==2.28.1
pytest==7.1.2
asyncpg==0.26.0
psycopg2==2.9.3
pydantic~=1.10.1